+++ packages/destinations/web/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @elbwalker/destination-web-api
 
+## 2.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @elbwalker/walker.js@3.0.1
+
 ## 2.3.0
 
 ### Minor Changes