+++ packages/destinations/web/google-ga4/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @elbwalker/destination-web-google-ga4
 
+## 3.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @elbwalker/walker.js@3.0.1
+
 ## 3.0.0
 
 ### Major Changes