+++ packages/destinations/web/piwikpro/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @elbwalker/destination-web-piwikpro
 
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @elbwalker/walker.js@3.0.1
+
 ## 1.2.0
 
 ### Minor Changes